long type applied

+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
 
         SalesManager newManager;
-        int[] weekSales = {2, 4, 7, 3, 5};
+        long[] weekSales = {2, 4, 7, 3, 5};
         newManager = new SalesManager(weekSales);
 
         System.out.println("newManager's week sales: " + Arrays.toString(newManager.sales));
